Move buildBrokerUrl to internal, keep it off the public surface
buildBrokerUrl produces core's private "/v1" STOMP broker path, so it is an
internal concern of StompConnectionManager — not public API. Living in
api/ConnectionInfo.ts leaked it onto the package entrypoint via `export *`,
which contradicts the goal of hiding the path from consumers.

Move it to internal/api/event/BrokerUrl.ts next to its primary caller.
StompConnectionManager and the public createAuthenticatedWebSocketFactory
both import it from there, so it stays the single source of truth while
consumers only ever touch the factory.
