+++ kinotic-js/workspace/packages/core/src/internal/api/event/StompConnectionManager.ts
@@ -1,4 +1,5 @@
-import {buildBrokerUrl, ConnectionInfo, type IWebSocket, SessionKeepAliveMode} from '@/api/ConnectionInfo'
+import {ConnectionInfo, type IWebSocket, SessionKeepAliveMode} from '@/api/ConnectionInfo'
+import {buildBrokerUrl} from '@/internal/api/event/BrokerUrl'
 import {EventConstants} from '@/api/event/IEventBus'
 import {ConnectedInfo} from '@/api/security/ConnectedInfo'
 import {type IFrame, RxStomp, RxStompConfig, StompHeaders} from '@stomp/rx-stomp'
